requirements.txt
langchain
langchain_core
langchain_community
streamlit
openai
python-dotenv
google-genai
langchain-openai
chromadb
beautifulsoup4
langchain-groq
PyPDF2